+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.13.1
 - Code coverage for internal simulator.
+- Simulator: A StepInto will now do a single step also for LDIR, LDDR, CPIR, CPDR and HALT.
 
 # 0.13.0
 - New command: "clearAllDecorations".